Rename TP_Python_Estadísitica_Actuarial.ipynb to Actuarial Statistics
The project marks the closure of the Actuarial Statistics course. The purpose is to articulate the theoretical concepts studied in the classes with a practical application using the programming language used in the course.

Throughout the course, different topics related to data analysis have been addressed. This project will seek to combine the skills learned in terms of programming and those analytical, logical, and problem-solving skills; with the ultimate goal of generating a framework that allows for a more comprehensive understanding of the subject matter.

The project is divided into three blocks that address different areas of statistical analysis:

1. Descriptive statistics and data management
2. Time series analysis
3. Simulation

The dataset is called WorldCupMatches and it contains data about all the football world cups made between 1930 and 2014.